cors now happens before other things. Closes #12.
